backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
python-dotenv==1.0.1
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
greenlet==3.1.1
# AI/LLM Dependencies (Member 3)
groq==0.4.1
google-generativeai==0.3.2
